+++ android/src/main/kotlin/com/adhub/adhub/AdhubPlugin.kt
@@ -1,0 +1,9 @@
+package com.adhub.adhub
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+class AdhubPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
+}