+++ ios/Classes/AdhubPlugin.swift
@@ -1,0 +1,5 @@
+import Flutter
+
+public class AdhubPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}